+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `reduce` function (`$set->reduce($callable, $initial)`)
+- Add `walk` function (`$set->walk($callable, $userdata)`)
+- Add `toJson` function (`$set->toJson($options, $depth)`)
+- Add `implode` function (`$set->implode($glue)`)
 
 ## [0.6.0] - 2019-11-21
 ### Changed